backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
PyPDF2
python-multipart
haystack-ai
elasticsearch
sentence-transformers
